refactor: 상수 데이터 파일 위치 변경 및 변수 이름 변경
